backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
asyncpg
bcrypt
pydantic[email]
pydantic-settings
python-multipart
Faker
requests
python-dateutil